Make it possible to serialize a Texture2D.
Meridian: New World relies on this.

XNA most likely allows this by accident, due to C++/CLI adding
default constructors to everything, and not having any properties
with setters. Meridian: New World doesn't seem to actually
serialize any textures, it just serializes a structure with a
texture field that seems to always be null when serialized.