Treat all keys as DialogKey instead of CmdKey
We previously treated all keys as a CmdKey to fix double characters when
entering text in Terraria.

Secrets of Grindea uses InputKey to trigger TranslateMessage, but
InputKey is called after CmdKey, so treating all keys as CmdKey breaks
it.

DialogKey is called after InputKey, so this fixes both Secrets of
Grindea and Terraria.